Prepare for public release: hardening, drivers, and consistent DominionOS naming
Security hardening (reviewed + adversarially verified, then fixed):
- 3 critical fixes: X.509 intermediate CA basicConstraints enforcement,
  BFT view-change lock authentication, forgeable fleet threshold signature.
- Dozens of high/medium fixes: unbounded allocations from untrusted input
  (NVMe/netif/object/persist/bft parsers), AES-GCM nonce reuse (memcrypt,
  memenc), spinlock-with-interrupts deadlocks (keyboard, mouse), ELF entry
  bounds check, DNS port binding, and more.
- Deferred protocol hardening now implemented: X25519 low-order-point
  rejection (RFC 7748), BPE tokenizer byte-level round-trip + map lookups,
  Proof-of-Learning Merkle step binding, identity moved to a hash-based
  Merkle one-time-signature scheme (verifier stores no secret), and an
  IPsec-style anti-replay window on session frames.

New hardware drivers:
- Intel e1000/e1000e Gigabit NIC (82540EM/82574L/I217/I211), unified with
  virtio-net behind a Nic trait so ARP/ICMP/UDP/DHCP work over any of them.
- USB mass storage (Bulk-Only Transport + SCSI over xHCI), registered as a
  block device alongside virtio-blk/AHCI/NVMe.

Naming + docs:
- Standardized the whole project on DominionOS / the Dominion language /
  DominionLink (removed the legacy Aether brand and the DominonOS typo).
- Reconciled HARDWARE.md/PROJECT_STATUS.md/FEATURES.md and friends with the
  real driver set; removed effort/line-count claims; rewrote the README.

dominion-core: 1724 tests pass. kernel/boot/bench build clean.

+++ dominion-core/src/atw.rs
@@ -161,7 +161,7 @@ pub struct CapturedFrame {
 
 impl CapturedFrame {
     pub fn new(width: u32, height: u32) -> Self {
-        let count = (width * height) as usize;
+        let count = width as usize * height as usize;
         Self {
             width,
             height,
@@ -173,14 +173,14 @@ impl CapturedFrame {
 
     #[inline]
     pub fn set_pixel(&mut self, x: u32, y: u32, color: u32, depth: f32) {
-        let idx = (y * self.width + x) as usize;
+        let idx = y as usize * self.width as usize + x as usize;
         self.pixels[idx] = color;
         self.depth[idx] = depth;
     }
 
     #[inline]
     pub fn get_pixel(&self, x: u32, y: u32) -> (u32, f32) {
-        let idx = (y * self.width + x) as usize;
+        let idx = y as usize * self.width as usize + x as usize;
         (self.pixels[idx], self.depth[idx])
     }
 }
@@ -238,7 +238,7 @@ impl Atw {
         fov_y_rad: f32,
         aspect: f32,
     ) -> Vec<u32> {
-        let count = (width * height) as usize;
+        let count = width as usize * height as usize;
         let mut out = alloc::vec![0u32; count];
 
         let frame = match &self.last_frame {
